+++ Content.Client/Humanoid/MarkingPicker.xaml.cs
@@ -25,7 +25,7 @@ public sealed partial class MarkingPicker : Control
 
     public Action<MarkingSet>? OnMarkingAdded;
     public Action<MarkingSet>? OnMarkingRemoved;
-    public Action<MarkingSet>? OnMarkingDataChanged; //Wayfarer/Coyote: OnMarkingColorChange to OnMarkingDataChanged
+    public Action<MarkingSet>? OnMarkingDataChanged; // Coyote: OnMarkingColorChange to OnMarkingDataChanged
     public Action<MarkingSet>? OnMarkingRankChange;
 
     private List<Color> _currentMarkingColors = new();
@@ -146,7 +146,7 @@ public sealed partial class MarkingPicker : Control
         CMarkingRankUp.OnPressed += _ => SwapMarkingUp();
         CMarkingRankDown.OnPressed += _ => SwapMarkingDown();
 
-        //Wayfarer/Coyote: Marking System Improvements start
+        // Coyote: Marking System Improvements start
         CanPutOnToggle.OnToggled += x => SetCanToggle(x.Pressed);
         CanPutOnByOtherToggle.OnToggled += x => SetOtherCanToggle(x.Pressed);
         StartVisibleToggle.OnToggled += x => SetVisible(x.Pressed);
@@ -256,13 +256,13 @@ public sealed partial class MarkingPicker : Control
     public void PopulateUsed()
     {
         CMarkingsUsed.Clear();
-        MarkingData.Visible = false; //Wayfarer/Coyote
+        MarkingData.Visible = false; // Coyote
         CMarkingColors.Visible = false;
         _selectedMarking = null;
 
-        SampleBox.Visible = false; //Wayfarer/Coyote
-        SampleButton.Visible = false; //Wayfarer/Coyote
-        SampleButton.Text = Loc.GetString("marking-show-sample-text"); //Wayfarer/Coyote
+        SampleBox.Visible = false; // Coyote
+        SampleButton.Visible = false; // Coyote
+        SampleButton.Text = Loc.GetString("marking-show-sample-text"); // Coyote
 
         if (!IgnoreSpecies)
         {
@@ -415,11 +415,11 @@ public sealed partial class MarkingPicker : Control
     {
         _selectedMarking = CMarkingsUsed[item.ItemIndex];
         var prototype = (MarkingPrototype) _selectedMarking.Metadata!;
-        // Wayfarer/Coyote: Marking System Improvements
+        // Coyote: Marking System Improvements
         int markingIndex = _currentMarkings.FindIndexOf(_selectedMarkingCategory, prototype.ID);
         if (markingIndex < 0) return;
         var marking = _currentMarkings.Markings[_selectedMarkingCategory][markingIndex];
-        // Wayfarer/Coyote End
+        // Coyote End
 
         if (prototype.ForcedColoring)
         {
@@ -493,7 +493,7 @@ public sealed partial class MarkingPicker : Control
             };
             colorSelector.OnColorChanged += colorChanged;
         }
-        // Wayfarer/Coyote Marking System Improvements
+        // Coyote Marking System Improvements
         CustomNameTextEdit.Text = marking.CustomName ?? "";
         StartVisibleToggle.Pressed = marking.ShowAtStart;
         CanPutOnToggle.Pressed = marking.CanToggleVisible;
@@ -507,7 +507,7 @@ public sealed partial class MarkingPicker : Control
         CMarkingColors.Visible = true; //Unchanged
         SampleButton.Visible = true;
         SetCheckboxVisibility();
-        // Wayfarer/Coyote end
+        // Coyote end
     }
 
     //Wayfarer: This is where the digi legs code would go, if we had it.
@@ -525,7 +525,7 @@ public sealed partial class MarkingPicker : Control
         marking.SetColor(colorIndex, _currentMarkingColors[colorIndex]);
         _currentMarkings.Replace(_selectedMarkingCategory, markingIndex, marking);
 
-        OnMarkingDataChanged?.Invoke(_currentMarkings); //Wayfarer/Coyote: OnMarkingColorChange? to OnMarkingDataChanged?
+        OnMarkingDataChanged?.Invoke(_currentMarkings); // Coyote: OnMarkingColorChange? to OnMarkingDataChanged?
     }
 
     private void MarkingAdd()
@@ -600,7 +600,7 @@ public sealed partial class MarkingPicker : Control
             Metadata = marking,
         };
         CMarkingsUsed.Insert(0, item);
-        OnUsedMarkingSelected(new ItemList.ItemListSelectedEventArgs(0, CMarkingsUsed)); //Wayfarer/Coyote: Select the newly added marking
+        OnUsedMarkingSelected(new ItemList.ItemListSelectedEventArgs(0, CMarkingsUsed)); // Coyote: Select the newly added marking
 
         _selectedUnusedMarking = null;
         OnMarkingAdded?.Invoke(_currentMarkings);
@@ -624,155 +624,8 @@ public sealed partial class MarkingPicker : Control
             item.Metadata = marking;
         }
         _selectedMarking = null;
-        MarkingData.Visible = false; // Wayfarer/Coyote
+        MarkingData.Visible = false; // Coyote
         CMarkingColors.Visible = false;
         OnMarkingRemoved?.Invoke(_currentMarkings);
     }
-
-    #region Wayfarer/Coyote
-    private void SetCheckboxVisibility()
-    {
-        if (CanPutOnToggle.Pressed)
-        {
-            PutOnTextEdit.Visible = true;
-            TakeOffTextEdit.Visible = true;
-            PutOnTextEditLabel.Visible = true;
-            TakeOffTextEditLabel.Visible = true;
-        }
-        else
-        {
-            PutOnTextEdit.Visible = false;
-            TakeOffTextEdit.Visible = false;
-            PutOnTextEditLabel.Visible = false;
-            TakeOffTextEditLabel.Visible = false;
-        }
-        if (CanPutOnByOtherToggle.Pressed)
-        {
-            PutOnOtherTextEdit.Visible = true;
-            TakeOffOtherTextEdit.Visible = true;
-            PutOnOtherTextEditLabel.Visible = true;
-            TakeOffOtherTextEditLabel.Visible = true;
-        }
-        else
-        {
-            PutOnOtherTextEdit.Visible = false;
-            TakeOffOtherTextEdit.Visible = false;
-            PutOnOtherTextEditLabel.Visible = false;
-            TakeOffOtherTextEditLabel.Visible = false;
-        }
-    }
-    private void SetCanToggle(bool canToggle)
-    {
-        if (_selectedMarking is null) return;
-        var markingPrototype = (MarkingPrototype)_selectedMarking.Metadata!;
-        int markingIndex = _currentMarkings.FindIndexOf(_selectedMarkingCategory, markingPrototype.ID);
-
-        if (markingIndex < 0) return;
-
-        var marking = new Marking(_currentMarkings.Markings[_selectedMarkingCategory][markingIndex]);
-        marking.CanToggleVisible = canToggle;
-        _currentMarkings.Replace(_selectedMarkingCategory, markingIndex, marking);
-        SetCheckboxVisibility();
-
-        OnMarkingDataChanged?.Invoke(_currentMarkings);
-    }
-
-    private void SetOtherCanToggle(bool canToggle)
-    {
-        if (_selectedMarking is null) return;
-        var markingPrototype = (MarkingPrototype)_selectedMarking.Metadata!;
-        int markingIndex = _currentMarkings.FindIndexOf(_selectedMarkingCategory, markingPrototype.ID);
-
-        if (markingIndex < 0) return;
-
-        var marking = new Marking(_currentMarkings.Markings[_selectedMarkingCategory][markingIndex]);
-        marking.OtherCanToggleVisible = canToggle;
-        _currentMarkings.Replace(_selectedMarkingCategory, markingIndex, marking);
-        SetCheckboxVisibility();
-
-        OnMarkingDataChanged?.Invoke(_currentMarkings);
-    }
-
-    private void SetVisible(bool visible)
-    {
-        if (_selectedMarking is null) return;
-        var markingPrototype = (MarkingPrototype)_selectedMarking.Metadata!;
-        int markingIndex = _currentMarkings.FindIndexOf(_selectedMarkingCategory, markingPrototype.ID);
-
-        if (markingIndex < 0) return;
-
-        var marking = new Marking(_currentMarkings.Markings[_selectedMarkingCategory][markingIndex]);
-        marking.ShowAtStart = visible;
-        _currentMarkings.Replace(_selectedMarkingCategory, markingIndex, marking);
-
-        OnMarkingDataChanged?.Invoke(_currentMarkings);
-    }
-
-    private void SetCustomText()
-    {
-        if (_selectedMarking is null) return;
-        var markingPrototype = (MarkingPrototype)_selectedMarking.Metadata!;
-        int markingIndex = _currentMarkings.FindIndexOf(_selectedMarkingCategory, markingPrototype.ID);
-
-        if (markingIndex < 0) return;
-
-        var marking = new Marking(_currentMarkings.Markings[_selectedMarkingCategory][markingIndex]);
-
-        marking.CustomName = CustomNameTextEdit.Text;
-        marking.PutOnVerb = PutOnTextEdit.Text;
-        marking.PutOnVerb2p = PutOnOtherTextEdit.Text;
-        marking.TakeOffVerb = TakeOffTextEdit.Text;
-        marking.TakeOffVerb2p = TakeOffOtherTextEdit.Text;
-
-        SampleText.Text = GetSampleText((string.IsNullOrEmpty(marking.CustomName) ? markingPrototype.ID : marking.CustomName),
-        (string.IsNullOrEmpty(marking.PutOnVerb) ? Loc.GetString("marking-toggle-self-default-verb-on") : marking.PutOnVerb),
-        (string.IsNullOrEmpty(marking.PutOnVerb2p) ? Loc.GetString("marking-toggle-other-default-verb-on") : marking.PutOnVerb2p))
-            + "\n" + GetSampleText((string.IsNullOrEmpty(marking.CustomName) ? markingPrototype.ID : marking.CustomName),
-        (string.IsNullOrEmpty(marking.TakeOffVerb) ? Loc.GetString("marking-toggle-self-default-verb-off") : marking.TakeOffVerb),
-        (string.IsNullOrEmpty(marking.TakeOffVerb2p) ? Loc.GetString("marking-toggle-other-default-verb-off") : marking.TakeOffVerb2p));
-
-        _currentMarkings.Replace(_selectedMarkingCategory, markingIndex, marking);
-
-        OnMarkingDataChanged?.Invoke(_currentMarkings);
-    }
-
-    private void ToggleSample()
-    {
-        if (SampleBox.Visible)
-        {
-            SampleBox.Visible = false;
-            SampleButton.Text = Loc.GetString("marking-show-sample-text");
-        }
-        else
-        {
-            SetCustomText();
-            SampleBox.Visible = true;
-            SampleButton.Text = Loc.GetString("marking-hide-sample-text");
-        }
-    }
-
-    private string GetSampleText(string name, string verb, string verb2p)
-    {
-        return Loc.GetString("marking-toggle-self-start",
-            ("marking-name", name),
-            ("verb", verb))
-            + "\n" + Loc.GetString("marking-toggle-self",
-            ("marking-name", name),
-            ("verb", verb))
-            + "\n" + Loc.GetString("marking-toggle-other-start",
-            ("marking-name", name),
-            ("verb", verb))
-            + "\n" + Loc.GetString("marking-toggle-other",
-            ("marking-name", name),
-            ("verb", verb))
-            + "\n" + Loc.GetString("marking-toggle-by-other-start",
-            ("other", "Someone"),
-            ("marking-name", name),
-            ("verb", verb))
-            + "\n" + Loc.GetString("marking-toggle-by-other",
-            ("other", "Someone"),
-            ("marking-name", name),
-            ("verb", verb2p));
-    }
-    #endregion Wayfarer/Coyote
 }
