+++ Content.Shared/Humanoid/Markings/MarkingsSet.cs
@@ -257,7 +257,7 @@ public sealed partial class MarkingSet
 
                 if (marking.Sprites.Count != list[i].MarkingColors.Count)
                 {
-                    list[i] = new Marking(list[i], marking.Sprites.Count); // Wayfarer/Coyote: marking.ID to list[i]
+                    list[i] = new Marking(list[i], marking.Sprites.Count); // Coyote: marking.ID to list[i]
                 }
             }
